requirements.txt
agno
openai
python-dotenv
os
ipykernel
groq
duckduckgo-search
tavily-python
